Move query building methods into query builder model
