+++ spec/factories/searches.rb
@@ -1,7 +1,7 @@
 FactoryBot.define do
   factory :search do
     keyword { "MyString" }
-    max_time { 1 }
-    allergies { "MyText" }
+    max_time { 35 }
+    allergies { "thing" }
   end
 end
