+++ spec/requests/searches/index_user_searches_spec.rb
@@ -1,0 +1,11 @@
+require 'rails_helper'
+
+describe 'GET /users/:token/searches' do
+
+  # it 'can access all users searches ordered by newest to oldest' do
+  # end
+  #
+  # it 'can access all users searches ordered oldest to newest' do
+  #
+  # end
+end
